% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_extraction.R
\name{combined_extraction}
\alias{combined_extraction}
\title{Estimate site terms by a combination of fixed effect and random effect models.}
\usage{
combined_extraction(
  data,
  path = FALSE,
  path_method = c("stepwise", "random", "fixed")
)
}
\arguments{
\item{data}{A data frame used for estimating site effect.}

\item{path}{A logical variable. If path == TRUE, path effect will be
controlled in estimation of site effect.}

\item{path_method}{A character string specifying the method to be used.
Options are "stepwise", "random", "fixed".}
}
\value{
A data frame that includes site ID, and the corresponding
true site terms and estimated site terms.
}
\description{
Estimate site terms by a combination of fixed effect and random effect models.
}
\examples{
data(gm_data)
combined_extraction(data = gm_data, path = TRUE, path_method = "random")
}
