% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_data.R
\docType{data}
\name{gm_data}
\alias{gm_data}
\title{Example Dataset}
\format{
A data frame with 200 rows and 6 variables:
\describe{
\item{r}{A numeric vector.}
\item{delta_B}{A numeric vector.}
\item{delta_S2S}{A numeric vector.}
\item{delta_WS}{A numeric vector.}
\item{event_id}{An integer vector.}
\item{site_id}{An integer vector.}
}
}
\usage{
gm_data
}
\description{
A sample ground motion dataset.
}
\examples{
data(gm_data)
head(gm_data)
}
\keyword{datasets}
