% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_extraction.R
\name{mixed_extraction}
\alias{mixed_extraction}
\title{Combined Extraction of Site Terms Using Fixed and Random Effect Models}
\usage{
mixed_extraction(
  data,
  path = FALSE,
  path_method = c("stepwise", "random", "fixed")
)
}
\arguments{
\item{data}{A data frame containing the input data for estimating site terms.
The data frame must include \code{site_id}, \code{event_id}, \code{r} (ground motion residuals), and optionally \code{path_id} if path effects are included.}

\item{path}{A logical variable indicating whether to control for path effects in the estimation.
\itemize{
\item \code{path = FALSE} (default): Path effects are not considered.
\item \code{path = TRUE}: Path effects are included, and the modeling strategy is determined by the \code{path_method} parameter.
}}

\item{path_method}{A character string specifying the method to control for path effects.
Options are \code{"stepwise"}, \code{"random"}, or \code{"fixed"}. This parameter is only required if \code{path = TRUE}.
\itemize{
\item "stepwise": Removes path effects stepwise.
\item "random": Treats path effects as random effects.
\item "fixed": Treats path effects as fixed effects.
}}
}
\value{
A data frame containing the site IDs (\code{site_id}) and their corresponding estimated site terms (\code{delta_S2S}).
}
\description{
The function estimate site terms by combining fixed effect and random effect models.
}
\examples{
data(gm_data)
mixed_extraction(data = gm_data)
}
