% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_extraction.R
\name{stepwise_extraction}
\alias{stepwise_extraction}
\title{Stepwise Extraction of Site Terms}
\usage{
stepwise_extraction(data, random = FALSE, path = FALSE)
}
\arguments{
\item{data}{A data frame containing the input data for estimating site terms.
The data frame must include columns \code{event_id}, \code{site_id}, \code{r} (ground motion residuals), and optionally \code{path_id} if path effects are included.}

\item{random}{A logical variable indicating whether to use random effect models.
\itemize{
\item \code{random = FALSE} (default) uses fixed effect models.
\item \code{random = TRUE} uses random effect models (based on \code{lme4::lmer()}).
}}

\item{path}{A logical variable indicating whether to account for path effects.
\itemize{
\item \code{path = FALSE} (default) estimates site terms without considering path effects.
\item \code{path = TRUE} removes path terms before estimating site terms.
}}
}
\value{
A data frame containing site IDs (\code{site_id}) and their corresponding estimated site terms (\code{delta_S2S}).
}
\description{
The function estimate site terms by removing the event terms (and path terms) from total ground motion residuals in a stepwise manner.
}
\examples{
data(gm_data)
delta_S2S <- stepwise_extraction(gm_data, random = FALSE)
}
