% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_extraction.R
\name{bayesian_extraction}
\alias{bayesian_extraction}
\title{Estimate site terms by Bayesian model.}
\usage{
bayesian_extraction(data, path = FALSE)
}
\arguments{
\item{data}{A data frame used for estimating site effect.}

\item{path}{A logical variable. If path == TRUE, path effect will be
controlled in estimation of site effect.}
}
\value{
A data frame that includes site ID, and the corresponding
true site terms and estimated site terms.
}
\description{
Estimate site terms by Bayesian model.
}
\examples{
data(gm_data)
bayesian_extraction(data = gm_data, path = FALSE)
}
