% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp}
\alias{dgp}
\title{Generate simulation data}
\usage{
dgp(
  num_eq,
  num_site,
  eta_eq_sd = 1,
  eta_site_sd = 1,
  eta_noise_sd = 1,
  path_term = FALSE,
  eta_path_sd = NULL,
  num_eq_cluster = NULL,
  num_site_cluster = NULL,
  missing_percent = 0
)
}
\arguments{
\item{num_eq}{The integer to specify the number of earthquakes.}

\item{num_site}{The integer to specify the number of earthquake stations.}

\item{eta_eq_sd}{The numeric value to specify the standard deviation of earthquake terms.}

\item{eta_site_sd}{The numeric value specify the standard deviation of station terms.}

\item{eta_noise_sd}{The numeric value specify the standard deviation of noise terms.}

\item{path_term}{A logical indicator for whether to generate path term.}

\item{eta_path_sd}{The numeric value to specify the standard deviation of path terms.}

\item{num_eq_cluster}{The integer to specify the number of earthquakes clusters.}

\item{num_site_cluster}{The integer to specify the number of station clusters.}

\item{missing_percent}{A decimal between 0 and 1, indicating the percentage of
missingness generated by missing completely at random(MCAR).}
}
\value{
A data frame containing the full simulation dataset.
}
\description{
Generate simulation data
}
\examples{
gm_data <- dgp(num_eq = 10, num_site = 20,
            eta_eq_sd = 1, eta_site_sd = 1, eta_noise_sd = 1,
            path_term = TRUE, eta_path_sd = 1,
            num_eq_cluster = 4, num_site_cluster = 5,
            missing_percent = 0.1)
}
