% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_data.R
\docType{data}
\name{gm_data}
\alias{gm_data}
\title{Example Dataset}
\format{
A data frame with 180 rows and 8 variables:
\describe{
\item{r}{A numeric vector.}
\item{eta_eq}{A numeric vector.}
\item{eta_site}{A numeric vector.}
\item{eta_path}{A numeric vector.}
\item{noise}{A numeric vector.}
\item{earthquake_id}{An integer vector.}
\item{site_id}{An integer vector.}
\item{path_id}{An integer vector.}
}
}
\usage{
gm_data
}
\description{
A sample ground motion dataset.
}
\examples{
data(gm_data)
head(gm_data)
}
\keyword{datasets}
