% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{remap_elements_and_levels}
\alias{remap_elements_and_levels}
\title{Remap Elements and Levels of a Factor Variable}
\usage{
remap_elements_and_levels(factor_var)
}
\arguments{
\item{factor_var}{A factor variable whose elements and levels are to be remapped.}
}
\value{
A factor variable with remapped elements and levels.
The new levels will be a sequential range starting from 1.
}
\description{
This function remaps the elements and levels of a factor variable.
Given a factor variable, it reassigns the levels to a sequential range
starting from 1 to the number of unique levels.
The factor elements themselves are also reassigned accordingly.
}
