% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_extraction.R
\name{stepwise_extraction}
\alias{stepwise_extraction}
\title{Estimate site terms by removing the earthquake effects, and path effects
if path is provided, using either fixed effect or random effect models.}
\usage{
stepwise_extraction(data, random = FALSE, path = FALSE)
}
\arguments{
\item{data}{A data frame used for estimating site effect.}

\item{random}{A logical variable. If random == TRUE, fixed effect models are used;
if random == FALSE, random effect models are used.}

\item{path}{A logical variable. If path == TRUE, path effect will be removed
in estimation of site effect.}
}
\value{
A data frame that includes site ID, and the corresponding
true site terms and estimated site terms.
}
\description{
Estimate site terms by removing the earthquake effects, and path effects
if path is provided, using either fixed effect or random effect models.
}
\examples{
data(gm_data)
eta_site_fitted <- stepwise_extraction(gm_data, random = FALSE, path = FALSE)
}
